requirements.txt
Flask
pip
flask_sqlalchemy
psycopg2
click
sreamlit
requests
airflow
spacy
jupyter
numpy
pandas
sklearn
flake8
en-core-web-lg
python-dotenv

#python -m spacy download en_core_web_lg